Attributes complete

+++ CourseProject/FileSizeManager.cpp
@@ -12,7 +12,7 @@ FileSizeManager::~FileSizeManager()
 }
 
 
-int FileSizeManager::getFileSize(BY_HANDLE_FILE_INFORMATION file)
+long FileSizeManager::getFileSize(BY_HANDLE_FILE_INFORMATION file)
 {
 	return file.nFileSizeLow;
 }
